implement rsg::Mailbox
This is an important milestone as Mailboxes are the first complex
object which are given a stub on the client side. I mean that
rsg::Mailbox lives only in the remote processes while s4u::Mailbox are
used in the rsg_server that runs the simulation.

I'm not a big fan of the current implementation, as rsg::Mailbox and
s4u::Mailbox do share a very large amount of code. Simply, s4u:: uses
simcall_* to get information while rsg:: uses networked requests.

Also, the code for networked requests should be placed in the engine,
not in the actor. The difficulties will come for multithreaded
applications, as I'm not sure of how to make the RSG state thread-local.

But at least it works and I can work further.