[code, test] rsg::this_actor::sleep_for
Test is known to fail on tesh, but I could not make it fail by calling
the commands manually so I still push this commit.

I felt that tesh was not suited for this kind of tests as fine-grained
synchronization between commands would enable reliable and fast tests,
but it should be enough for this test — as long as the machine that runs
the tests is not slow enough to break the ugly sleep-based sync written
in the test definition file).

The problem may be related to the client termination, as I had similar
problems with tesh since the quit message has been introduced.
Solving https://framagit.org/simgrid/simgrid/issues/23 would help there.