[code] serve: signal via fd, signal->TCP RST
First is to keep logic straightforward.
No more signal callback that exits.

Second is to make sure that the server finishes fast when it receives a
signal (SIGINT, SIGTERM, SIGSEGV). This is by dropping any pending
message on all the sockets (in the two directions) on close.
This is done via SO_LINGER, that resets the connection when used with
the parameters used here.