rsg: fix condvar use (temporary unique_locks)
Temporary unique_locks were created in the RSG server whenever
a wait function was called remotely. This caused lock ownership
inconsistency, as the destructor of this temporary unique_lock
called unlock() on the mutex, which not happened on the remote mutex.

This fix creates a single unique_lock per mutex (with the same lifecycle
of its mutex), and uses this unique_lock in all condvar wait functions.