code: enable remote actor think time simulation
Goal is to make it simple, in a rsg simulation, to simulate the remote
process thinking time, that is to say the amount of real time it used between
receiving a DECISION_ACK and sending the next DECISION.

This first implementation only supports a constant time increase,
e.g., all decisions from a remote actor will take X seconds in the simulation.
If X=0, decision taking time is not simulated.

Improving this version so that each remote client measures its decision taking time
should be quite simple:
- enable a custom string in rsg add-actor --think-time, forward it in RSG_THINK_TIME
- take time point measures in librsg in Connection::send_decision,
  inject it in the protobuf message